+++ src/api/__tests__/api.test.js
@@ -1,5 +1,8 @@
 import getData from '..';
 import { request } from '../helpers';
+import {
+  VEHICLES_RESOURCE, PORT, HOST, API_DIR
+} from '../config';
 
 jest.mock('../helpers');
 
@@ -23,17 +26,17 @@ describe('getData Tests', () => {
     request.mockResolvedValueOnce([]);
     await safelyCallApi();
 
-    expect(request).toBeCalledWith('/api/vehicles.json');
+    expect(request).toBeCalledWith(`${HOST}${PORT}${API_DIR}${VEHICLES_RESOURCE}`);
   });
 
   it('Should traverse and make further api calls on main results', async () => {
     expect.assertions(3);
-    request.mockResolvedValueOnce([{ apiUrl: '/api/vehicle_ftype.json' }, { apiUrl: '/api/vehicle_xj.json' }]);
+    request.mockResolvedValueOnce([{ id: 'ftype', apiUrl: '/api/vehicle_ftype.json' }, { id: 'xj', apiUrl: '/api/vehicle_xj.json' }]);
     request.mockResolvedValueOnce({ id: 'ftype', price: '£36,000' });
     request.mockResolvedValueOnce({ id: 'xj', price: '£40,000' });
     await safelyCallApi();
 
-    expect(request).toBeCalledWith('/api/vehicles.json');
+    expect(request).toBeCalledWith(`${HOST}${PORT}${API_DIR}${VEHICLES_RESOURCE}`);
     expect(request).toBeCalledWith('/api/vehicle_ftype.json');
     expect(request).toBeCalledWith('/api/vehicle_xj.json');
   });
@@ -48,7 +51,7 @@ describe('getData Tests', () => {
   });
 
   it('Should ignore vehicles without valid price during traversing', () => {
-    request.mockResolvedValueOnce([{ apiUrl: '/api/ftype.json' }, { apiUrl: '/api/xe.json' }, { apiUrl: '/api/xj.json' }]);
+    request.mockResolvedValueOnce([{ id: 'ftype', apiUrl: '/api/ftype.json' }, { id: 'xe', apiUrl: '/api/xe.json' }, { id: 'xj', apiUrl: '/api/xj.json' }]);
     request.mockResolvedValueOnce({ id: 'ftype', price: '' });
     request.mockResolvedValueOnce({ id: 'xe' });
     request.mockResolvedValueOnce({ id: 'xj', price: '£40,000' });